+++ index.html
@@ -250,7 +250,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script defer src="/js/main.js"></script>
+  <script defer type="module" src="/js/main.js"></script>
 </body>
 
 </html>
